address cell conversion error
In rare cases of Ibn/Icn modeling, the CSM cells can have radius inversion that fail to construct the CSM profile at tinj.
We address this by removing the problematic cell(s) and interpolating the profile with the sane cells
We thank Craig Pellegrino for spotting and letting us know this error.